[DRAFT] merge odf_table element and internal representation
Idea:

class odf_table(odf_element):
  ...

So the table becomes an element class like the others. It stores the
original XML tree of the table. Repetitions and the like for limited
memory use.

But...

It also contains an internal state, empty by default.

You can ask for a cell, a row, a column (and later a region). It will
return the value in the state or the default value from the XML tree.

When setting a cell, the internal state is filled. Blank positions
before the coordinates of the cell are filled with None.

The returned cells are writable but only taken into account when put
back in the state through the API.

Only the required cells will be expanded to memory.

Expanding a row and its cells will obviously have a cost. But I think
the memory save is worth it.

When serializing the table, internal structure is synchronized with the
XML tree and repetitions are applied.

When the document is saved, all tables must be synchronized.